Tape and robot (#2258)
* Add queue support to workflow

* Fix tape access and locking

* Update workspace with second dir level

* Add db migrations

* Update requirements

* Fix tests and package information view

* Updated workflow for access to workspace

* Fix DIP filebrowser to workspace

* Fix expand of container files

* Fix add file to DIP from other workspaces

+++ ESSArch_Core/ip/signals.py
@@ -53,9 +53,10 @@ def workarea_post_delete(sender, instance, using, **kwargs):
         if e.errno != errno.ENOENT:
             raise
 
-    if not Workarea.objects.filter(ip__aic=instance.ip.aic).exists():
-        try:
-            os.remove(instance.aic_xml_path)
-        except OSError as e:
-            if e.errno != errno.ENOENT:
-                raise
+    if instance.ip.aic is not None:
+        if not Workarea.objects.filter(ip__aic=instance.ip.aic).exists():
+            try:
+                os.remove(instance.aic_xml_path)
+            except OSError as e:
+                if e.errno != errno.ENOENT:
+                    raise
